# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OleoControl",
        "description": "Documentación de la API de la aplicación OleoControl",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Listar todos los empleados",
                "description": "Muestra todos los empleados.\n\nEste método obtiene todos los empleados de la base de datos y devuelve una respuesta JSON con un estado de éxito y los datos de los empleados.",
                "operationId": "5f4bcd5a07bb6c8d34f2feaf68f3a7c7",
                "responses": {
                    "200": {
                        "description": "Lista de empleados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EmployeeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crear un nuevo empleado",
                "description": "Crea un nuevo empleado junto con el usuario y rol correspondiente.\n\nEste método crea un nuevo empleado, su usuario asociado y le asigna el rol de \"Empleado\". La creación se realiza dentro de una transacción para garantizar la integridad de los datos.\nLa respuesta incluye un estado de éxito y los datos del usuario creado.",
                "operationId": "7fc7d604ba7e8873bab5fcf3839b6d90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "dni",
                                    "email",
                                    "phone",
                                    "status",
                                    "department_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Laura"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Martínez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678Z"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "laura@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "666555444"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "activo"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Mostrar un empleado",
                "description": "Muestra un empleado específico por su ID.\n\nEste método recibe un ID de empleado, busca el empleado en la base de datos y devuelve una respuesta JSON con un estado de éxito y los datos del empleado.",
                "operationId": "69cd01de9e0b2cd120d584ef3e639f96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del empleado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EmployeeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar un empleado",
                "description": "Actualiza un empleado específico por su ID.\n\nEste método recibe una solicitud de actualización de empleado, valida los datos y actualiza el empleado y su usuario asociado en la base de datos.\nLa respuesta incluye un estado de éxito y los datos del empleado actualizado en formato JSON.",
                "operationId": "b22e1fd0d650fa5c63c8a7d1d32acbfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user": {
                                        "properties": {
                                            "username": {
                                                "type": "string",
                                                "example": "laura.martinezZ"
                                            },
                                            "first_name": {
                                                "type": "string",
                                                "example": "Laura"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Martínez"
                                            },
                                            "dni": {
                                                "type": "string",
                                                "example": "12345678Z"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "laura@example.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "666555444"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "activo"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "department": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/EmployeeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Eliminar un empleado",
                "description": "Elimina un empleado específico por su ID.\n\nEste método recibe un ID de empleado, busca el empleado en la base de datos y elimina el empleado.\nLa respuesta incluye un estado de éxito y un mensaje de éxito en formato JSON.",
                "operationId": "9a570eeea95ac9960cc5cfbeccc26123",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario del empleado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Empleado eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmployeeResource": {
                "title": "EmployeeResource",
                "description": "Recurso que representa un empleado, incluyendo su usuario y el departamento al que pertenece",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "username": {
                                "type": "string",
                                "example": "johndoe"
                            },
                            "first_name": {
                                "type": "string",
                                "example": "John"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "Doe"
                            },
                            "dni": {
                                "type": "string",
                                "example": "12345678A"
                            },
                            "email": {
                                "type": "string",
                                "example": "johndoe@example.com"
                            },
                            "phone": {
                                "type": "string",
                                "example": "123456789"
                            },
                            "status": {
                                "type": "string",
                                "example": "active"
                            },
                            "roles": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Administrator"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "department": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "HR"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticación Bearer (usado por Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}